requirements.txt
python-telegram-bot==20.0a0
requests==2.27
fake-useragent==1.1.3
